requirements.txt
flask
requests
pillow
gunicorn
pytesseract
gevent
influxdb
